a11y: consistent naming in english (notification -> report), fix title thankyou page
