+++ src/app/[locale]/incident/thankyou/page.tsx
@@ -1,40 +1,23 @@
-'use client'
-
-import { useTranslations } from 'next-intl'
-import { Paragraph, Heading } from '@/components/index'
-import { Button, PreserveData } from '@utrecht/component-library-react'
 import React from 'react'
-import { stepToPath, useRouter } from '@/routing/navigation'
-import { FormStep } from '@/types/form'
-import { useFormStore } from '@/store/form_store'
+import { Metadata } from 'next/types'
+import { getServerConfig } from '@/services/config/config'
+import { getTranslations } from 'next-intl/server'
+import { createTitle } from '@/lib/utils/create-title'
+import { ThankyouContent } from '@/app/[locale]/incident/thankyou/components/ThankyouContent'
+
+export async function generateMetadata(): Promise<Metadata> {
+  const config = await getServerConfig()
+  const t = await getTranslations('describe_thankyou')
+  const tGeneral = await getTranslations('general.form')
+
+  return {
+    title: createTitle(
+      [t('heading'), config.base.naam],
+      tGeneral('title_separator')
+    ),
+  }
+}
 
 export default function Thankyou() {
-  const t = useTranslations('describe_thankyou')
-  const { resetForm, formState } = useFormStore()
-  const router = useRouter()
-
-  const resetState = () => {
-    resetForm()
-    router.push(stepToPath[FormStep.STEP_1_DESCRIPTION])
-  }
-
-  return (
-    <div className="flex flex-col gap-8">
-      <Heading level={1}>{t('heading')}</Heading>
-      <div className="flex flex-col gap-2">
-        <Paragraph>
-          {t('description_notification_number')}
-          <PreserveData> {formState.sig_number}</PreserveData>
-        </Paragraph>
-        <Paragraph>{t('description_notification_email')}</Paragraph>
-      </div>
-      <div className="flex flex-col gap-2">
-        <Heading level={3}>{t('what_do_we_do_heading')}</Heading>
-        <Paragraph>{t('what_do_we_do_description')}</Paragraph>
-      </div>
-      <Button appearance="primary-action-button" onClick={() => resetState()}>
-        {t('new_notification')}
-      </Button>
-    </div>
-  )
+  return <ThankyouContent />
 }
