+++ translations/nl.json
@@ -89,7 +89,7 @@
     }
   },
   "describe_thankyou": {
-    "heading": "Bedankt!",
+    "heading": "Melding verzonden",
     "description_notification_number": "Uw melding is bij ons bekend onder nummer:",
     "description_notification_email": "Hebt u een e-mailadres ingevuld? Dan ontvangt u een e-mail met alle gegevens van uw melding.",
     "what_do_we_do_heading": "Wat doen we met uw melding?",
